Move IActiveSolutionBoundTracker to Core
