+++ src/Integration/Telemetry/TelemetryManager.cs
@@ -24,6 +24,7 @@ using System.Diagnostics;
 using System.Linq;
 using Microsoft.VisualStudio.Shell;
 using SonarLint.VisualStudio.Core;
+using SonarLint.VisualStudio.Core.Binding;
 using SonarLint.VisualStudio.Core.SystemAbstractions;
 using SonarLint.VisualStudio.Core.Telemetry;
